Fixed daily DatabaseCleanupService, daily cleanup runs midnight

+++ src/main/java/com/bloomboard/promptboard/security/service/DatabaseCleanupService.java
@@ -3,17 +3,17 @@ package com.bloomboard.promptboard.security.service;
 import com.bloomboard.promptboard.prompt.Prompt;
 import com.bloomboard.promptboard.prompt.PromptService;
 import com.bloomboard.promptboard.security.model.User;
-import com.bloomboard.promptboard.security.model.UserRole;
+import com.bloomboard.promptboard.tag.TagService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Service
+@Component
 @RequiredArgsConstructor
 public class DatabaseCleanupService {
 
@@ -23,18 +23,38 @@ public class DatabaseCleanupService {
     private final UserServiceImpl userService;
     @Autowired
     private final PromptService promptService;
+    @Autowired
+    private final TagService tagService;
 
-    @Scheduled(cron = "0 0 0 * * ?") // Run every day at midnight
+    //@Scheduled(fixedRate = 5000) // Schedule to run every 5 seconds
+    public void testScheduledMethod() {
+        logger.info("TEST MESSAGE SENT!!!!");
+    }
+
+    @Scheduled(cron = "0 0 0 * * ?") //Everynight at midnight
     public void performDailyCleanup() {
-        UserRole demoRole = UserRole.DEMO;
-        List<User> demoUsers = userService.findByUserRole(demoRole);
+        try {
+            logger.info("Performing daily cleanup...");
 
-        for ( User user: demoUsers) {
-            List<Prompt> demoPrompts = promptService.findByUser_id(user.getId());
-            for (Prompt prompt: demoPrompts) {
-                promptService.deletePrompt(prompt.getId());
+            //Delete posts from demo account
+            User demoUser = userService.findByUsernameIgnoreCase("DemoUser");
+            List<Prompt> demoPrompts = promptService.findByUser_id(demoUser.getId());
+            if(!demoPrompts.isEmpty()) {
+                for (Prompt prompt: demoPrompts) {
+                    promptService.deletePrompt(prompt.getId());
+                }
+                logger.info("\tDeleted " + demoPrompts.size() + " prompts from the DemoUser account.");
+                logger.info("\tDeleted Prompts: \n\t\t" + demoPrompts);
+            } else {
+                logger.info("\tNo prompts deleted from DemoUser account");
             }
-        }
 
+            //delete tags not associated with a prompt
+            tagService.deleteOrphanedTags();
+
+            logger.info("Daily cleanup completed successfully.");
+        } catch (Exception e){
+            logger.error("An error occurred during daily cleanup:", e);
+        }
     }
 }
